chore: create i18n module
to load TranslateModule.root() with correct config
