+++ libs/common/src/index.ts
@@ -4,3 +4,4 @@
 
 export * from './lib/common.service'
 export * from './lib/common.module'
+export * from './lib/i18n/i18n.module'
