+++ libs/common/src/lib/i18n/i18n.module.ts
@@ -1,0 +1,33 @@
+import { HttpClient, HttpClientModule } from '@angular/common/http'
+import { NgModule } from '@angular/core'
+import {
+  TranslateLoader,
+  TranslateModule,
+  TranslateService,
+} from '@ngx-translate/core'
+import { TranslateHttpLoader } from '@ngx-translate/http-loader'
+
+export function HttpLoaderFactory(http: HttpClient) {
+  return new TranslateHttpLoader(http)
+}
+
+@NgModule({
+  declarations: [],
+  imports: [
+    HttpClientModule,
+    TranslateModule.forRoot({
+      loader: {
+        provide: TranslateLoader,
+        useFactory: HttpLoaderFactory,
+        deps: [HttpClient],
+      },
+    }),
+  ],
+  exports: [TranslateModule],
+})
+export class I18nModule {
+  constructor(translate: TranslateService) {
+    translate.setDefaultLang('en')
+    translate.use('en')
+  }
+}
